Moved contents of feedback_config directory into project_admin.
