+++ tests/test_project_admin/test_ag_suites/test_ag_test_command_advanced_fdbk_settings.ts
@@ -2,7 +2,7 @@ import { config, mount, Wrapper } from '@vue/test-utils';
 
 import { AGTestCase, AGTestCommandFeedbackConfig, AGTestSuite, ValueFeedbackLevel } from 'ag-client-typescript';
 
-import EditFeedbackSettingsAGCommand from '@/components/feedback_config/edit_feedback_settings/edit_feedback_settings_ag_command.vue';
+import EditFeedbackSettingsAGCommand from '@/components/project_admin/ag_suites/ag_test_command_advanced_fdbk_settings.vue';
 
 import * as data_ut from '@/tests/data_utils';
 import { checkbox_is_checked } from '@/tests/utils';
